Simplified code
Moved files around to avoid code duplication.
