--- src/components/ResultsSearchSection/ResultsSearchSection.css
@@ -1,26 +1,0 @@
-.results-search-container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 300px;
-    background-color: #00000010;
-    padding-left: 5%;
-}
-
-.results-search-container h1{
-    text-align: left;
-    margin-bottom: 5px;
-    margin-top: 80px;
-}
-
-.results-search-container .searchandresults {
-    display: block;
-    align-items: center;
-    justify-content: center;
-    background-color: #ffffff;
-    border: solid #00000015 2px;
-    border-radius: 30px;
-    width: fit-content;
-    /* box-shadow: 0px 2px #0000001b; */
-    z-index: 1;
-}
